完成： 1)增加将shopID通过时间戳的方式设置成唯一ID 2)修正了addShopInfo 和 queryShopInfos的存储逻辑 待解决： 1）addShopInfo 需要解决一个用户添加多个商店信息的问题

+++ src/main/java/education/cs/scu/entity/PropertyBean.java
@@ -10,8 +10,8 @@ import java.util.List;
 public class PropertyBean implements Serializable {
 
     private static final long serialVersionUID = 351877796426921776L;
-    private int propertyId;
-    private int shopId;
+    private Integer propertyId;
+    private Long shopId;
     private String mmac;
     private String visitCycle;
     private double visitRange;
@@ -27,7 +27,7 @@ public class PropertyBean implements Serializable {
         return serialVersionUID;
     }
 
-    public int getPropertyId() {
+    public Integer getPropertyId() {
         return propertyId;
     }
 
@@ -35,11 +35,11 @@ public class PropertyBean implements Serializable {
         this.propertyId = propertyId;
     }
 
-    public int getShopId() {
+    public long getShopId() {
         return shopId;
     }
 
-    public void setShopId(int shopId) {
+    public void setShopId(long shopId) {
         this.shopId = shopId;
     }
